feature admin delete image

+++ src/commons/ProductModal.jsx
@@ -14,31 +14,33 @@ import {
   ordenProductosCarrusel,
 } from "../assets/auxiliarFunctions";
 import { useSelector } from "react-redux";
+import { useNavigate } from "react-router-dom";
+import DeleteModal from "./DeleteModal";
+import { toast } from "react-toastify";
 
 function ProductModal({ modalOpen, handleModal }) {
+  const navigate = useNavigate();
   const adminLogged = useSelector(
     (state) => state.user.userData?.payload.is_admin
   );
   const [productInfo, setProductInfo] = useState({});
   const [productsCarrusel, setProductCarrusel] = useState([]);
-  const [move, setMove] = useState(false);
   const [index, setIndex] = useState(0);
 
   const carruselMoveRight = () => {
-    setMove(true);
+   
     if (index < productsCarrusel.length - 1) {
       setIndex(index + 1);
     }
   };
   const carruselMoveLeft = () => {
-    setMove(true);
+   
     if (index > 0) {
       setIndex(index - 1);
     }
   };
 
   const id = useLocation().search.split("=")[1];
-
 
   useEffect(() => {
     const fetchData = async () => {
@@ -80,7 +82,7 @@ function ProductModal({ modalOpen, handleModal }) {
             id
           );
           setProductCarrusel(productosConImagenesOrdenados);
-          setIndex(0)
+          setIndex(0);
         }
       } catch (error) {
         console.error(error);
@@ -90,29 +92,76 @@ function ProductModal({ modalOpen, handleModal }) {
     fetchData();
   }, [id]);
 
-  console.log("vista individual", productInfo);
-  console.log("carrusel", productsCarrusel);
+  const [modalDelete, setModalDelete] = useState(false);
+  const handleModalDelete = () => {
+    setModalDelete(!modalDelete);
+  };
 
   const handleDelete = () => {
-    return axios.delete(
-      `http://localhost:4000/api/v1/product/erase/${productsCarrusel[index]._id}`
-    );
+    if (productsCarrusel.length > 1) {
+      let newIndex;
+      if (index + 1 === productsCarrusel.length) {
+        newIndex = index - 1;
+      } else {
+        newIndex = index + 1;
+      }
+      axios
+        .delete(
+          `http://localhost:4000/api/v1/product/erase/${productsCarrusel[index]._id}`
+        )
+        .then(() => {
+          axios.delete(
+            `http://localhost:4000/api/v1/images/${productsCarrusel[index].url_img}`
+          );
+        })
+        .then(() => {
+          navigate(`?id=${productsCarrusel[newIndex]._id}`);
+        });
+    }
+    if (productsCarrusel.length === 1) {
+      axios
+        .delete(
+          `http://localhost:4000/api/v1/product/erase/${productsCarrusel[index]._id}`
+        )
+        .then(() => {
+          axios.delete(
+            `http://localhost:4000/api/v1/images/${productsCarrusel[index].url_img}`
+          );
+        })
+        .then(() => {
+          navigate("/rubros");
+          handleModal();
+        });
+    }
+    handleModalDelete();
+    toast.success("Producto Eliminado");
   };
 
   return (
     <>
       {modalOpen && productInfo.name && (
         <>
-          <div className="bg-[#f2f2f2] h-[90vh] w-[40vw] fixed top-[5%] left-[30%] z-40 rounded-md">
-            <div
-              onClick={() => {
-                handleModal();
-                setProductCarrusel([])
-              }}
-              className="w-full flex justify-end"
-            >
+          <DeleteModal
+            handleModalDelete={handleModalDelete}
+            modalDelete={modalDelete}
+            handleDelete={handleDelete}
+          />
+          <div
+            className={`${
+              modalDelete
+                ? "opacity-50 brightness-50 pointer-events-none bg-black "
+                : "bg-[#f2f2f2] "
+            }h-[90vh] w-[40vw] fixed top-[5%] left-[30%] z-40 rounded-md`}
+          >
+            <div className="w-full flex justify-end">
               <Link to={`/rubros`}>
-                <FaTimes style={{ marginRight: "1%", fontSize: "1.5rem" }} />
+                <FaTimes
+                  onClick={() => {
+                    handleModal();
+                    setProductCarrusel([]);
+                  }}
+                  style={{ marginRight: "1%", fontSize: "1.5rem" }}
+                />
               </Link>
             </div>
 
@@ -123,14 +172,8 @@ function ProductModal({ modalOpen, handleModal }) {
                 </p>
                 {adminLogged && (
                   <FaTrash
-                    onClick={handleDelete}
-                    style={{
-                      position: "absolute",
-                      right: "6%",
-                      fontSize: "1.4rem",
-                      color: "red",
-                      marginTop: "2%",
-                    }}
+                    onClick={handleModalDelete}
+                    className="tachitoBasura"
                   />
                 )}
                 <img
@@ -140,14 +183,18 @@ function ProductModal({ modalOpen, handleModal }) {
               </div>
             )}
           </div>
-          <FaChevronLeft
-            onClick={carruselMoveLeft}
-            className="leftArrowCarrusel"
-          />
-          <FaChevronRight
-            onClick={carruselMoveRight}
-            className="rightArrowCarrusel"
-          />
+          {!modalDelete && (
+            <>
+              <FaChevronLeft
+                onClick={carruselMoveLeft}
+                className="leftArrowCarrusel"
+              />
+              <FaChevronRight
+                onClick={carruselMoveRight}
+                className="rightArrowCarrusel"
+              />
+            </>
+          )}
         </>
       )}
     </>
